+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 All notable changes to the "nc-view" extension will be documented in this file.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how
+to structure this file.
 
 ## [Unreleased]
 
-- Initial release
+### Added
+
+- v 0.0.4 - Add support uppercase file extensions and exclude codes setting